Kernel: Move Lockable into its own header
