+++ Kernel/Net/LocalSocket.cpp
@@ -9,6 +9,8 @@
 #include <Kernel/Debug.h>
 #include <Kernel/FileSystem/FileDescription.h>
 #include <Kernel/FileSystem/VirtualFileSystem.h>
+#include <Kernel/Locking/Lockable.h>
+#include <Kernel/Locking/Mutex.h>
 #include <Kernel/Net/LocalSocket.h>
 #include <Kernel/Process.h>
 #include <Kernel/StdLib.h>
