+++ Kernel/Net/UDPSocket.cpp
@@ -6,6 +6,7 @@
 
 #include <AK/Singleton.h>
 #include <Kernel/Devices/RandomDevice.h>
+#include <Kernel/Locking/Mutex.h>
 #include <Kernel/Net/NetworkAdapter.h>
 #include <Kernel/Net/Routing.h>
 #include <Kernel/Net/UDP.h>
